🔨 move focus array to its own folder
